+++ molecule/config.py
@@ -252,7 +252,6 @@ class Config(object):
                 'enabled': True,
                 'options': {},
                 'env': {},
-                'trailing_ignore_paths': ['.git', '.vagrant', '.molecule']
             },
             'platforms': [],
             'provisioner': {